♻️ (AddNicknameForm): move to forms dir
