+++ src/components/Forms/AddNicknameForm/AddNicknameForm.module.scss
@@ -1,5 +1,5 @@
-@import '../../scss/animations/toast-fade';
-@import '../../scss/colors';
+@import '../../../scss/animations/toast-fade';
+@import '../../../scss/colors';
 
 .addNicknameForm {
   display: flex;
